added the suggested changes from the reviewers
