+++ lib/components/footer/bottomNavigationTabs.dart
@@ -2,8 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:off_top_mobile/login.dart';
 import 'package:off_top_mobile/recordingSession.dart';
 import 'package:off_top_mobile/settings_page.dart';
-
-import '../../reports.dart';
+import 'package:off_top_mobile/reports.dart';
 
 class BottomNavigationTabs extends StatelessWidget {
   const BottomNavigationTabs(this.homePage);
